refactor: lazy load cron subagent followup runtime
