+++ src/cron/isolated-agent/subagent-followup.runtime.ts
@@ -1,0 +1,4 @@
+export {
+  readDescendantSubagentFallbackReply,
+  waitForDescendantSubagentSummary,
+} from "./subagent-followup.js";
